Realtime version passthrough (#282)
Pass through the realtime version string to the `RecoInterface` object
in order to give reconstructions the ability to determine what keys to
use.

For example, for pointed scans, the particle around which to scan used
to be a list of particles names that contained variations across all
realtime versions. This is brittle and prone to unexpected behavior. By
passing the version string to `RecoInterface` instances the appropriate
key can be chosen based on that.

Additionally, remove `INPUT_PULSES_NAME` in config.py in favor of
accessing it via a static method in the RecoInterface via the
`INPUT_KEYS_MAP` defined in config.py. This removes the need to copy /
rename pulseseries in the frame, and also allows for future key name
changes with the realtime format.

Changes a few base class variables that were later shadowed by instance
variables into instance variables in the base class itself.

---------

Co-authored-by: wipacdevbot <developers@icecube.wisc.edu>